feat: 기능 추가
- 수동으로 구매할 로또 수 입력 기능 및 문구 추가
- 수동으로 구매할 로또 번호 입력 기능 및 문구 추가
- 수동으로 구매한 로또 수에 따라 자동으로 구매할 수 있는 개수 계산 기능 추가
-
